requirements.txt
Flask==1.1.2
openpyxl==3.0.6
pandas==1.2.1
pyodbc==4.0.30
Flask-SQLAlchemy==2.4.4